Split core hooks module
